Added all element
